D2: decompose channel_cog into views/channels/* package
Reduces cogs/channel_cog.py from 1497 LOC to 464 LOC (69% reduction)
by extracting every panel view into a dedicated module under
disbot/views/channels/. The cog file now hosts only the prefix
commands and channel-resolution helpers; everything UI-related lives
in the views package.

New package layout:
  disbot/views/channels/
  ├── __init__.py           re-exports (58 LOC)
  ├── _helpers.py           presets + _build_channel_options +
  │                         _ChannelSelect (used by delete + restrict)
  ├── main_panel.py         _ChannelManagerView (top-level hub)
  ├── create_panel.py       _CreateSubView + _NameSelect +
  │                         _CategorySelect + _CustomNameModal
  ├── delete_panel.py       _DeleteSubView + _DeleteConfirmView
  ├── restrict_panel.py     _RestrictSubView (lock/unlock flow)
  └── visibility_panel.py   _VisibilitySubView +
                            _ChannelSelectForVisibility +
                            _SubsystemToggleView (per-channel toggles)

Persistence rule honoured: no SUBSYSTEM string, no custom_id, and no
panel_anchors.subsystem value was renamed. The visibility toggle's
``custom_id="toggle_<subsystem>"`` is unchanged. The cog is not
persistent-view-backed (every view is ephemeral, BaseView-derived)
so no panel anchors are involved.

Cross-panel back/cancel navigation uses local imports inside button
callbacks (lazy import pattern) to break the circular module graph
between main_panel and the four sub-panels.

Back-compat re-exports: cogs/channel_cog.py re-imports every panel
class from the new package, so existing imports like
``from cogs.channel_cog import _SubsystemToggleView`` keep resolving.

Tests:
- tests/unit/views/test_view_error_handling.py — repointed two patch
  targets from ``cogs.channel_cog.governance_service.…`` to
  ``views.channels.visibility_panel.governance_service.…`` since the
  governance import now lives in the panel module. The view-class
  import itself remains unchanged thanks to the re-exports.

Verification: ruff/black/isort/mypy clean; pytest 436/436 green.
